chore: update samples and add tests
