+++ spanner/spanner_enable_fine_grained_access.rb
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# [START spanner_list_database_roles]
+# [START spanner_enable_fine_grained_access]
 require "google/cloud/spanner"
 
-def spanner_list_database_roles project_id:, instance_id:, database_id:, iam_member:, database_role:, title:
+def spanner_enable_fine_grained_access project_id:, instance_id:, database_id:, iam_member:, database_role:, title:
   # project_id  = "Your Google Cloud project ID"
   # instance_id = "Your Spanner instance ID"
   # database_id = "Your Spanner database ID"
@@ -26,7 +26,7 @@ def spanner_list_database_roles project_id:, instance_id:, database_id:, iam_mem
   admin_client = Google::Cloud::Spanner::Admin::Database::V1::DatabaseAdmin::Client.new
   db_path = admin_client.database_path project: project_id, instance: instance_id, database: database_id
 
-  policy = admin_client.get_iam_policy
+  policy = admin_client.get_iam_policy resource: db_path, options: { requested_policy_version: 3 }
 
   policy.version = 3 if policy.version < 3
 
@@ -40,29 +40,12 @@ def spanner_list_database_roles project_id:, instance_id:, database_id:, iam_mem
   )
     
   policy.bindings << binding
-  admin_client.set_iam_policy resource: db_path, policy: policy
+  result = admin_client.set_iam_policy resource: db_path, policy: policy
  
-   new_binding = policy_pb2.Binding(
-       role="roles/spanner.fineGrainedAccessUser",
-       members=[iam_member],
-       condition=expr_pb2.Expr(
-           title=title,
-           expression=f'resource.name.endsWith("/databaseRoles/{database_role}")',
-       ),
-   )
- 
-   policy.version = 3
-   policy.bindings.append(new_binding)
-   database.set_iam_policy(policy)
- 
-   new_policy = database.get_iam_policy(3)
-   print(
-       f"Enabled fine-grained access in IAM. New policy has version {new_policy.version}"
-   )
-
+  puts "Enabled fine-grained access in IAM."
 end
-# [END spanner_list_database_roles]
+# [END spanner_enable_fine_grained_access]
 
 if $PROGRAM_NAME == __FILE__
-    spanner_list_database_roles project_id: ARGV.shift, instance_id: ARGV.shift, database_id: ARGV.shift
+  spanner_enable_fine_grained_access project_id: ARGV.shift, instance_id: ARGV.shift, database_id: ARGV.shift
 end
