+++ spanner/spec/spec_helper.rb
@@ -16,6 +16,7 @@ require "rspec"
 require "google/cloud/spanner"
 require "google/cloud/spanner/admin/instance"
 require "google/cloud/spanner/admin/database"
+require_relative "../spanner_samples"
 
 RSpec.configure do |config|
   config.before :all do
@@ -126,5 +127,111 @@ RSpec.configure do |config|
 
     job.wait_until_done!
     database = job.results
+  end
+
+    # Creates a temporary database with random ID (will be dropped after test)
+  # (re-uses create_database to create database with Albums/Singers schema)
+  def create_singers_albums_database
+    capture do
+      create_database project_id:  @project_id,
+                      instance_id: @instance.instance_id,
+                      database_id: @database_id
+
+      @test_database = @instance.database @database_id
+    end
+
+    @test_database
+  end
+
+  def create_performances_table
+    capture do
+      create_table_with_timestamp_column project_id:  @project_id,
+                                         instance_id: @instance.instance_id,
+                                         database_id: @database_id
+    end
+  end
+
+  def create_venues_table
+    capture do
+      create_table_with_datatypes project_id:  @project_id,
+                                  instance_id: @instance.instance_id,
+                                  database_id: @database_id
+    end
+  end
+
+  def create_boxes_database
+    job = @instance.create_database @database_id
+    job.wait_until_done!
+    @test_database = job.database
+  end
+
+  def create_database_with_data
+    database = create_singers_albums_database
+
+    capture do
+      write_using_dml project_id:  @project_id,
+                      instance_id: @instance.instance_id,
+                      database_id: database.database_id
+
+      @test_database = @instance.database @database_id
+    end
+
+    @test_database
+  end
+
+  # Creates or return existing temporary backup with random ID (will be dropped
+  # after test)
+  def create_backup_with_data
+    @test_backup = @instance.backup @backup_id
+
+    return @test_backup if @test_backup
+
+    database = create_singers_albums_database
+
+    capture do
+      write_using_dml project_id:  @project_id,
+                      instance_id: @instance.instance_id,
+                      database_id: database.database_id
+    end
+
+    client = @spanner.client @instance.instance_id, database.database_id
+    version_time = client.execute("SELECT CURRENT_TIMESTAMP() as timestamp").rows.first[:timestamp]
+
+    capture do
+      create_backup project_id:   @project_id,
+                    instance_id:  @instance.instance_id,
+                    database_id:  database.database_id,
+                    backup_id:    @backup_id,
+                    version_time: version_time
+
+      @test_backup = @instance.backup @backup_id
+    end
+
+    @test_backup
+  end
+
+  def create_copy_backup
+    capture do
+      copy_backup project_id: @project_id,
+                  instance_id: @instance.instance_id,
+                  backup_id: @copied_backup_id,
+                  source_backup_id: @backup_id
+    end
+    @instance.backup @copied_backup_id
+  end
+
+  def restore_database_from_backup
+    backup = create_backup_with_data
+
+    capture do
+      restore_backup project_id:  @project_id,
+                     instance_id: @instance.instance_id,
+                     database_id: @restored_database_id,
+                     backup_id:   backup.backup_id
+
+      @test_database = @instance.database @restored_database_id
+    end
+
+    @test_database
   end
 end
